Improve performance of Pwned using binary search
The previous implementation used the index to locate the ~50 or so
hashes with the same prefix as the search target, but then simply
scanned through them all sequentially until the target was found.
This was very fast if the target was early in the range and became
progressively slower for targets nearer the end; absent targets
scanned through the whole range before being rejected.

Now a binary search algorithm is used to narrow the search space in
`log_2(N)` time, cutting the number of comparisons down from ~50 to
just 5 or 6. This radically reduces the cost of an individual check
for obvious reasons. For a hot hash, the check latency now reaches
as low as 60 nanoseconds!

    name      old time/op  new time/op  delta
    Pwned-10   924ns ± 3%    59ns ± 4%  -93.62%  (p=0.000 n=20+20)