Improve performance of pwngen
Evidently using `binary.Write` is very slow? It has been replaced with
direct use of `dff`. I found that this improved index generation throughput
from ~250 MB/s to ~500 MB/s on a 2021 M1 MacBook Pro.

This inefficiency was masked because the passwords database is distributed
compressed with 7z and decompressing it was already the bottleneck. Using
an uncompressed copy of the archive (or one recompressed with Zstandard) it
is now possible to generate the index in about a minute.